+++ examples/backdround_with_image.rb
@@ -1,0 +1,5 @@
+Shoes.app do
+  stack width:200, height:400 do
+    background "docs/static/avatar.png"
+  end
+end
