+++ lib/scarpe/wv/image.rb
@@ -1,14 +1,16 @@
 # frozen_string_literal: true
 
-require "base64"
-require "uri"
+# require "base64"
+# require "uri"
+require_relative "../base64"
 
 class Scarpe
   class WebviewImage < WebviewWidget
+    include Base64
     def initialize(properties)
       super
 
-      @url = valid_url?(@url) ? @url : "data:image/png;base64,#{encode_image_to_base64(@url)}"
+      @url = valid_url?(@url) ? @url : "data:image/png;base64,#{encode_file_to_base64(@url)}"
     end
 
     def element
@@ -24,25 +26,6 @@ class Scarpe
     end
 
     private
-
-    def valid_url?(string)
-      uri = URI.parse(string)
-      uri.is_a?(URI::HTTP) || uri.is_a?(URI::HTTPS)
-    rescue URI::InvalidURIError, URI::BadURIError
-      false
-    end
-
-    def encode_image_to_base64(image_filename)
-      directory_path = File.dirname(__FILE__, 4)
-
-      image_path = File.join(directory_path, image_filename)
-
-      image_data = File.binread(image_path)
-
-      encoded_data = Base64.strict_encode64(image_data)
-
-      encoded_data
-    end
 
     def style
       styles = {}
